- Cleanup/typo on homunc's code - Modified @reloadmobdb & @reloadskilldb to reload homunc's db too
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@8476 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/map.h
@@ -986,7 +986,7 @@ struct homun_data {
 	struct map_session_data *master; //pointer back to its master
 	int hungry_timer;	//[orn]
 	int target_id,attacked_id;
-	unsigned long exp_next;
+	unsigned int exp_next;
 	char blockskill[MAX_SKILL];	// [orn]
 };
